move _helper into the tests subpackage
